Specification Adjustments
